refactor(common): move PR schemas + helpers to kolu-common/pr subpath
Runtime-importing `prValue` / `prUnavailableReason` from the root
`kolu-common` entry transitively pulled `@anthropic-ai/claude-agent-sdk`
into the browser bundle (via kolu-claude-code's re-export chain), and the
SDK references Node built-ins — breaking `vite build`. A dedicated
`./pr.ts` module keeps the helpers reachable without dragging the agent
SDK along; clients now import from `kolu-common/pr`.
